Moved declarations to seperate headers. Reorganised the directory structure to better suit the interpreters structure.
