+++ src/lib/include/curlyc.h
@@ -1,0 +1,11 @@
+#ifndef CURLY_C_H
+#define CURLY_C_H
+
+#include <builtins.h>
+#include <io.h>
+#include <lval.h>
+#include <lenv.h>
+#include <macros.h>
+#include <windows.h>
+
+#endif  /// CURLY_C_H
