+++ src/lib/include/types.h
@@ -1,0 +1,13 @@
+#ifndef CURLY_TYPES_H
+#define CURLY_TYPES_H
+
+/// \brief Returns the name of a type
+///
+/// \details Returns a string of of the passed
+/// type's name.
+///
+/// \param t - type int
+/// \return char*
+char* ltype_name(int t);
+
+#endif  /// CURLY_TYPES_H
