+++ src/lib/src/lenv.c
@@ -1,0 +1,8 @@
+#include <lenv.h>
+
+struct lenv 
+{
+    int count;
+    char** syms;
+    lval** vals;
+};
