+++ src/lib/src/lval.c
@@ -1,0 +1,19 @@
+#include <lval.h>
+
+struct lval
+{
+    int type;
+
+    long num;
+    char* err;
+    char* sym;
+
+    lbuiltin builtin;
+    lenv* env;
+    lval* formals;
+    lval* body;
+
+    int count;
+    struct lval** cell;
+};
+
