+++ src/lib/src/parsing.c
@@ -1,73 +1,5 @@
 #include <parsing.h>
 
-#define LASSERT(args, cond, fmt, ...)               \
-    if (!(cond))                                    \
-    {                                               \
-        lval* err = lval_err(fmt, ##__VA_ARGS__);   \
-        lval_del(args);                             \
-        return err;                                 \
-    }
-
-
-#define LASSERT_TYPE(func, args, index, expect)                             \
-  LASSERT(args, args->cell[index]->type == expect,                          \
-    "Function '%s' passed incorrect type for argument %i. "                 \
-    "Got %s, Expected %s.",                                                 \
-    func, index, ltype_name(args->cell[index]->type), ltype_name(expect))
-
-
-#define LASSERT_NUM(func, args, num)                                        \
-  LASSERT(args, args->count == num,                                         \
-    "Function '%s' passed incorrect number of arguments. "                  \
-    "Got %i, Expected %i.",                                                 \
-    func, args->count, num)
-
-#define LASSERT_NOT_EMPTY(func, args, index)                                \
-  LASSERT(args, args->cell[index]->count != 0,                              \
-    "Function '%s' passed {} for argument %i.", func, index);
-
-
-#ifdef _WIN32
-
-    static char buffer[2048];
-
-    char* readline(char* prompt) 
-    {
-        fputs(prompt, stdout);
-        fgets(buffer, 2048, stdin);
-        char* cpy = malloc(strlen(buffer) + 1);
-        strcpy(cpy, buffer);
-        cpy[strlen(cpy)-1] = '\0';
-        return cpy;
-    }
-
-    void add_history(char* unused) {}
-#endif
-
-struct lval
-{
-    int type;
-
-    long num;
-    char* err;
-    char* sym;
-
-    lbuiltin builtin;
-    lenv* env;
-    lval* formals;
-    lval* body;
-
-    int count;
-    struct lval** cell;
-};
-
-struct lenv 
-{
-    int count;
-    char** syms;
-    lval** vals;
-};
-
 lval* lval_num(long x)
 {
     lval* v = malloc(sizeof(lval));
